+++ src/main/java/com/hotdeal/reservation/payment/client/MockPgClient.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Component;
 public class MockPgClient implements PgClient {
 
     @Override
-    public void charge(int amount) {
+    public void charge(long amount) {
     }
 }
