+++ src/main/java/com/hotdeal/reservation/payment/client/PgClient.java
@@ -2,5 +2,5 @@ package com.hotdeal.reservation.payment.client;
 
 public interface PgClient {
 
-    void charge(int amount);
+    void charge(long amount);
 }
